Add LMDB upgrade to v22: Remove unchecked table
In pull request #4021 the binary representation of `unchecked_info` was
changed. We have to clear the unchecked table, because the new binary
format is not compatible with the old one.

This commit only clears the unchecked table for the LMDB implementation.
The RocksDB implementation doesn't have an upgrade mechanism yet.

- Delete the unchecked DB from the LMDB environment
- Simplify the unchecked handle