Migrate tests from the old message_parser to test the message_deserializer class (#4166)
* Replace message_parser test for confirm_ack to test message_deserializer

The old test checked the bytes generated the same confirm_ack object by
ensuring its visitor was processed only once. Also verified the
parse_status was success. In the new implementation the visitor isn't
part of the deserializer. The outcome object is compared to the
original, and the parse_status should be success. This is the sanity
test, where the deserialization is expected to happen successfully.

Also checks if the message pointer can be downcast to the proper type,
that is the same of verifying its type is supported.

* Move message_parser confirm_req tests to the message_deserializer format

* Move message_parser publish test to the message_deserializer format

* Move message_parser keepalive test to the message_deserializer format

* Remove the dev_visitor class

* Add message_deserialize tests for all the remaining message types

* Add more comments to the message_deserializer_success_checker function

* Update the file name to core_test/message_deserializer.cpp