Fix active_transactions.confirm_frontier
The main problem was that the bootstrap processes could sync the block
before the test tried to insert it and it woudl fail because the block
was already there. The main problem I fixed was to disable the bootstrap
processes. I also improved it overall and made it more readable.