Fix rpc.receivable_unconfirmed
In RPC speak, cofnirmed is equivalent to cemented and in this case the
test only check if the election is not active, which indirectly but not
directly implies that the election is confirmed. But even then it is not
necesarily cemented since that is an extra background step.

We should check for cemented with the function nano::test:confirmed(),
which is also confusing...