Move wallet receiving in to its own observer.
This is an interaction between the ledger when a block is confirmed and the wallets that might need to generate receive blocks. Previously this functionality was unnecessarily being handled by the active_transactions class. It could also miss notifications if the block was indirectly confirmed without an active election, or if the election had already been removed from active_transactions.