Use memory weight instead of exact weight.
This function is used to check if an account has enough weight to be a rep. If this account isn't in memory because of its high weight, by definition it doesn't have enough weight.
This also eliminates a database transaction while holding the reps_cache_mutex in nano::wallets::compute_reps. This mutex is in several hot paths of code such as nano::wallets::reps being called from message_visitor::confirm_req